+++ kwcoco/cli/coco_conform.py
@@ -53,7 +53,6 @@ class CocoConformCLI:
             >>> cls.main(cmdline, **kw)
         """
         import kwcoco
-        import kwimage
 
         config = cls.CLIConfig(kw, cmdline=cmdline)
         print('config = {}'.format(ub.repr2(dict(config), nl=1)))
@@ -65,25 +64,28 @@ class CocoConformCLI:
 
         dset = kwcoco.CocoDataset.coerce(config['src'])
 
-        if config['ensure_imgsize']:
-            dset._ensure_imgsize(workers=config['workers'])
+        config_ = ub.dict_diff(config, {'src', 'dst'})
+        dset.conform(**config_)
 
-        if config['pycocotools_info']:
-            for ann in ub.ProgIter(dset.dataset['annotations'], desc='update anns'):
-                if 'iscrowd' not in ann:
-                    ann['iscrowd'] = False
+        # if config['ensure_imgsize']:
+        #     dset._ensure_imgsize(workers=config['workers'])
 
-                if 'ignore' not in ann:
-                    ann['ignore'] = ann.get('weight', 1.0) < .5
+        # if config['pycocotools_info']:
+        #     for ann in ub.ProgIter(dset.dataset['annotations'], desc='update anns'):
+        #         if 'iscrowd' not in ann:
+        #             ann['iscrowd'] = False
 
-                if 'area' not in ann:
-                    # Use segmentation if available
-                    if 'segmentation' in ann:
-                        poly = kwimage.Polygon.from_coco(ann['segmentation'][0])
-                        ann['area'] = float(poly.to_shapely().area)
-                    else:
-                        x, y, w, h = ann['bbox']
-                        ann['area'] = w * h
+        #         if 'ignore' not in ann:
+        #             ann['ignore'] = ann.get('weight', 1.0) < .5
+
+        #         if 'area' not in ann:
+        #             # Use segmentation if available
+        #             if 'segmentation' in ann:
+        #                 poly = kwimage.Polygon.from_coco(ann['segmentation'][0])
+        #                 ann['area'] = float(poly.to_shapely().area)
+        #             else:
+        #                 x, y, w, h = ann['bbox']
+        #                 ann['area'] = w * h
 
         dset.fpath = config['dst']
         print('dump dset.fpath = {!r}'.format(dset.fpath))